kernel-build.eclass: install dtb into /lib/modules instead of /boot
Systemd's kernel-install hooks expect it here or in the firmware directory. We
need to find it for building UKI's and registering BLS type 1 entries. Fedora
and Arch also install the dtb files here.

If it is needed in /boot then it is probably a better idea to delegate that
task to /sbin/installkernel anyway. That way the part of the emerge that writes
to /boot is contained to only the postinst and config phases and makes it
possible to re-try this via emerge --config ... .

Closes: https://bugs.gentoo.org/945072
Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/39486
Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>